+++ src/Tasks/LegalProposition/Helpers.hs
@@ -3,7 +3,7 @@ module Tasks.LegalProposition.Helpers where
 import Tasks.SynTree.Config (SynTreeConfig (..))
 import Trees.Types (BinOp)
 
-formulaAmount :: SynTreeConfig -> Int
+formulaAmount :: SynTreeConfig -> Integer
 formulaAmount SynTreeConfig{..} =
-  (if allowArrowOperators then length [minBound..maxBound :: BinOp] else 2)
+  (if allowArrowOperators then fromIntegral (length [minBound..maxBound :: BinOp]) else 2)
     ^ ((maxNodes - 1) `div` 2 - minUniqueBinOperators + 1)
